operation_map supports exception handlers, add --version command
This cleans up the large try/catch structure in the cli by allowing
the operation map to define what happens when a certain exception
is encountered.

Because most operations in most cases will return
a dict, Error subclasses dict so that callers can check the return
value without relying on a particular structure for an Error.
This would also be possible by raising a new subclass of Exception,
but in most cases callers will do something with a dict-based structure
and just want and out-of-band signal for Error/Ok.